Move the CreateTenantCommand to a separate folder
