+++ account-management/tests/Tests/Application/Tenants/Commands/CreateTenantCommandHandlerTests.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using NSubstitute;
-using PlatformPlatform.AccountManagement.Application.Tenants.Commands;
+using PlatformPlatform.AccountManagement.Application.Tenants.Commands.CreateTenant;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
 using Xunit;
